requirements.txt
boto3
python-dotenv
pyaudio
SpeechRecognition
flask
playwright
pyttsx3